+++ R/track.R
@@ -1,4 +1,4 @@
-.track <- function(x, y, z, t, id, proj, origin, period, tz) {
+.track <- function(x, y, z, t, id, proj, origin, period, tz, format) {
   if (!is.numeric(x))
     stop("x must be an object of class numeric.")
 
@@ -13,11 +13,16 @@
   if (!lubridate::is.POSIXct(t)) {
     if (is.numeric(t)) {
       if (!missing(origin)) {
-        origin <- tryCatch(lubridate::as_datetime(origin, tz = tz), error = function(e) {
-          e$call <- "track_df(..., origin)"
-          e$message <- "Invalid origin format."
-          stop(e)
-        })
+        origin <- tryCatch(
+          if (missing(format))
+            lubridate::as_datetime(origin, tz = tz)
+          else
+            lubridate::as_datetime(origin, tz = tz, format = format),
+          error = function(e) {
+            e$call <- "track_df(..., origin)"
+            e$message <- "Invalid origin format."
+            stop(e)
+          })
       } else {
         message("No origin provided. Defaulting to Sys.time().")
         origin <- Sys.time()
@@ -34,12 +39,21 @@
         period <- lubridate::period("1 second")
       }
 
-      t <- lubridate::as_datetime(t * period, origin = origin, tz = tz)
+      t <- if (missing(format))
+        lubridate::as_datetime(t * period, origin = origin, tz = tz)
+      else
+        lubridate::as_datetime(t * period, origin = origin, tz = tz, format = format)
     } else {
-      t <- lubridate::as_datetime(t, tz = tz)
+      t <- if (missing(format))
+        lubridate::as_datetime(t, tz = tz)
+      else
+        lubridate::as_datetime(t, tz = tz, format = format)
     }
   } else {
-    t <- lubridate::as_datetime(t, tz = tz)
+    t <- if (missing(format))
+      lubridate::as_datetime(t, tz = tz)
+    else
+      lubridate::as_datetime(t, tz = tz, format = format)
   }
 
   if (!missing(id)) {
@@ -99,7 +113,7 @@
 #'  one.
 #'
 #' @param t A numeric vector or a vector of objects that can be coerced to
-#'  date-time objects by \code{link[lubridate]{as_datetime}} representing the
+#'  date-time objects by \code{\link[lubridate]{as_datetime}} representing the
 #'  times (or frames) at which each location was recorded. If numeric, the
 #'  origin and period of the time points can be set using \code{origin} and
 #'  \code{period} below.
@@ -120,7 +134,7 @@
 #'  \code{"+proj=longlat"} is suitable for the output of most GPS trackers.
 #'
 #' @param origin Something that can be coerced to a date-time object by
-#'  \code{link[lubridate]{as_datetime}} representing the start date and time of
+#'  \code{\link[lubridate]{as_datetime}} representing the start date and time of
 #'  the observations when \code{t} is a numeric vector.
 #'
 #' @param period A character vector in a shorthand format (e.g. "1 second") or
@@ -132,6 +146,11 @@
 #'  \code{\link[lubridate]{period}} for more details.
 #'
 #' @param tz A time zone name. See \code{\link{OlsonNames}}.
+#'
+#' @param format A character string indicating the formatting of `t`. See
+#'  \code{\link{strptime}} for how to specify this parameter.
+#'
+#' When supplied parsing is performed by strptime(). For this reason consider using specialized parsing functions in lubridate.
 #'
 #' @param table A string indicating the class of the table on which the track
 #'  table should be built. It can be a \code{\link{data.frame}} ("df", the default),
@@ -161,23 +180,23 @@
 #' @rdname track_
 #'
 #' @export
-track <- function(x, y, z, t, id, ..., proj, origin, period, tz, table = "df") {
+track <- function(x, y, z, t, id, ..., proj, origin, period, tz, format, table = "df") {
   switch(table,
-    "df" = track_df(x = x, y = y, z = z, t = t, id = id, ..., proj = proj,
-                    origin = origin, period = period, tz = tz),
-    "tbl" = track_tbl(x = x, y = y, z = z, t = t, id = id, ..., proj = proj,
-                      origin = origin, period = period, tz = tz),
-    "dt" = track_dt(x = x, y = y, z = z, t = t, id = id, ..., proj = proj,
-                    origin = origin, period = period, tz = tz),
-    stop("Unknown table type.")
+         "df" = track_df(x = x, y = y, z = z, t = t, id = id, ..., proj = proj,
+                         origin = origin, period = period, tz = tz, format = format),
+         "tbl" = track_tbl(x = x, y = y, z = z, t = t, id = id, ..., proj = proj,
+                           origin = origin, period = period, tz = tz, format = format),
+         "dt" = track_dt(x = x, y = y, z = z, t = t, id = id, ..., proj = proj,
+                         origin = origin, period = period, tz = tz, format = format),
+         stop("Unknown table type.")
   )
 }
 
 #' @rdname track_
 #'
 #' @export
-track_df <- function(x, y, z, t, id, ..., proj, origin, period, tz) {
-  l <- .track(x, y, z, t, id, proj, origin, period, tz)
+track_df <- function(x, y, z, t, id, ..., proj, origin, period, tz, format) {
+  l <- .track(x, y, z, t, id, proj, origin, period, tz, format)
   out <- as.data.frame(l[names(l) != "proj"], stringsAsFactors = FALSE)
 
   args <- list(...)
@@ -196,8 +215,8 @@ track_df <- function(x, y, z, t, id, ..., proj, origin, period, tz) {
 #' @rdname track_
 #'
 #' @export
-track_tbl <- function(x, y, z, t, id, ..., proj, origin, period, tz) {
-  l <- .track(x, y, z, t, id, proj, origin, period, tz)
+track_tbl <- function(x, y, z, t, id, ..., proj, origin, period, tz, format) {
+  l <- .track(x, y, z, t, id, proj, origin, period, tz, format)
   out <- tibble::as_tibble(l[names(l) != "proj"])
 
   args <- list(...)
@@ -216,8 +235,8 @@ track_tbl <- function(x, y, z, t, id, ..., proj, origin, period, tz) {
 #' @rdname track_
 #'
 #' @export
-track_dt <- function(x, y, z, t, id, ..., proj, origin, period, tz) {
-  l <- .track(x, y, z, t, id, proj, origin, period, tz)
+track_dt <- function(x, y, z, t, id, ..., proj, origin, period, tz, format) {
+  l <- .track(x, y, z, t, id, proj, origin, period, tz, format)
   out <- data.table::as.data.table(l[names(l) != "proj"])
 
   args <- list(...)